# Use sync.Pool in Listener's readLoop