<!--
PLEASE READ THIS MESSAGE.

Documentation:
- for Traefik v2: use branch v2.11 (fixes only)
- for Traefik v3: use branch v3.5

Bug:
- for Traefik v2: use branch v2.11 (security fixes only)
- for Traefik v3: use branch v3.5

Enhancements:
- use branch master

HOW TO WRITE A GOOD PULL REQUEST? https://doc.traefik.io/traefik/contributing/submitting-pull-requests/

-->

### What does this PR do?

<!-- A brief description of the change being made with this pull request. -->


### Motivation
Avoid unneeded allocations in `readLoop` by using `sync.Pool`
Verdict:
- Speed: sync-pool is 6.27% faster (9.51µs → 8.92µs, p=0.000).
- Memory: ~40% fewer bytes/op (64.14KiB → 38.73KiB, p=0.000).
- Allocs: goes from 6 → 7 allocs/op (+1). Despite that, total bytes drop a lot.

```
$benchstat old.txt sync-pool.txt                             
goos: darwin
goarch: arm64
pkg: github.com/traefik/traefik/v3/pkg/udp
cpu: Apple M3 Max
                    │   old.txt   │           sync-pool.txt            │
                    │   sec/op    │   sec/op     vs base               │
ReadLoopAllocations   9.511µ ± 2%   8.915µ ± 1%  -6.27% (p=0.000 n=20)

                    │   old.txt    │            sync-pool.txt             │
                    │     B/op     │     B/op      vs base                │
ReadLoopAllocations   64.14Ki ± 0%   38.73Ki ± 2%  -39.61% (p=0.000 n=20)

                    │  old.txt   │           sync-pool.txt            │
                    │ allocs/op  │ allocs/op   vs base                │
ReadLoopAllocations   6.000 ± 0%   7.000 ± 0%  +16.67% (p=0.000 n=20)
```


I run the benchmark with these flags
```
go test -run=^$ -bench=BenchmarkReadLoopAllocations -benchmem -cpu=1 -count=20
```

### More

- [ ] Added/updated tests
- [ ] Added/updated documentation

### Additional Notes

<!-- Anything else we should know when reviewing? -->