@rtribotte 2025-09-22
Hi @arturmelanchyk,

Sorry for the late reply and thanks this contribution!

The changes require careful review, which is why we labelled it as a design review.
Unfortunately, considering our limited bandwidth and the risk/opportunity balance (we are unaware of memory issues with the current design), it could take time to manage the review.
That being said, we understand the enhancement you are proposing, but we are wondering about the effects. For instance, how will this change impact the throughput? Have you already conducted any benchmarks in this regard?