@kevinpollet 2025-11-13
Hello @arturmelanchyk,

Just so you know, we reverted the changes of https://github.com/traefik/traefik/pull/12267, as it seems to have introduced a regression as reported in https://github.com/traefik/traefik/issues/12251.
We will look into it later.